API: Disable XML Service Responses
Our primary output format is JSON and we'd like everyone to use it. If
we could support XML for free that'd be OK, but because of the way we're
setting up DTOs (using inheritance and the like) XML schemas and
responses are basically impossible. Therefore it seems best to just
outright remove the XML formatter since it can't work.