API: Update Timeseries to use the correct Template logic
Correctly use the Template identifier, after converted to a master id,
when looking up timeseries and their components. We want to limit the
items returned on a direct timeseries GET to only those that are a
part of the template, and similar when a general query is made we
should only being returning those timeseries with a value for an item
for our template.