API: Break out SF Specific Cell Members
Some contents of the CellExpressionNode were clearly SF specific, and since
we're trying to build a common API that can be used by various different
efforts it makes sense to break out these things into a SF specific cell node.

The only issue here is that there isn't an easy way to convey via the API that
"SF should only ever use SFCellExpressionNode in their Expressions, and not
the original CellExpressionNode". The good news is that the actual API doesn't
care one bit, but I'm interested in how best to support this on the JS side of
things.